# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amane API",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/actors": {
      "get": {
        "tags": [
          "actors"
        ],
        "summary": "List Actors",
        "description": "分页列出演员 (人物摘要 + 关联影片数).",
        "operationId": "list_actors",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ActorSortField",
              "default": "name"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "has_person",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否已有人物标量/简介",
              "title": "Has Person"
            },
            "description": "是否已有人物标量/简介"
          },
          {
            "name": "has_image",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否已有头像 URL",
              "title": "Has Image"
            },
            "description": "是否已有头像 URL"
          },
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActorGender"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "性别多选 (可重复)",
              "title": "Gender"
            },
            "description": "性别多选 (可重复)"
          },
          {
            "name": "birthday_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "生日下界 YYYY-MM-DD",
              "title": "Birthday Min"
            },
            "description": "生日下界 YYYY-MM-DD"
          },
          {
            "name": "birthday_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "生日上界 YYYY-MM-DD",
              "title": "Birthday Max"
            },
            "description": "生日上界 YYYY-MM-DD"
          },
          {
            "name": "height_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "身高下界 cm",
              "title": "Height Min"
            },
            "description": "身高下界 cm"
          },
          {
            "name": "height_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "身高上界 cm",
              "title": "Height Max"
            },
            "description": "身高上界 cm"
          },
          {
            "name": "bust_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "胸围下界 cm",
              "title": "Bust Min"
            },
            "description": "胸围下界 cm"
          },
          {
            "name": "bust_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "胸围上界 cm",
              "title": "Bust Max"
            },
            "description": "胸围上界 cm"
          },
          {
            "name": "waist_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "腰围下界 cm",
              "title": "Waist Min"
            },
            "description": "腰围下界 cm"
          },
          {
            "name": "waist_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "腰围上界 cm",
              "title": "Waist Max"
            },
            "description": "腰围上界 cm"
          },
          {
            "name": "hip_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "臀围下界 cm",
              "title": "Hip Min"
            },
            "description": "臀围下界 cm"
          },
          {
            "name": "hip_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "臀围上界 cm",
              "title": "Hip Max"
            },
            "description": "臀围上界 cm"
          },
          {
            "name": "cup_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "罩杯下界 (字母序, 大小写不敏感)",
              "title": "Cup Min"
            },
            "description": "罩杯下界 (字母序, 大小写不敏感)"
          },
          {
            "name": "cup_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "罩杯上界",
              "title": "Cup Max"
            },
            "description": "罩杯上界"
          },
          {
            "name": "birthplace",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "出生地包含匹配",
              "title": "Birthplace"
            },
            "description": "出生地包含匹配"
          },
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "限制为这些演员主键",
              "title": "Ids"
            },
            "description": "限制为这些演员主键"
          },
          {
            "name": "saved_query_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Saved query preset id; AND with other filters via SQL subquery",
              "title": "Saved Query Id"
            },
            "description": "Saved query preset id; AND with other filters via SQL subquery"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/actors/{actor_id}": {
      "get": {
        "tags": [
          "actors"
        ],
        "summary": "Get Actor",
        "description": "演员详情 (含别名与 raw).",
        "operationId": "get_actor",
        "parameters": [
          {
            "name": "actor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Actor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "actors"
        ],
        "summary": "Update Actor",
        "description": "更新演员规范人物字段 (不含 name/raw/field_sources).",
        "operationId": "update_actor",
        "parameters": [
          {
            "name": "actor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Actor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActorUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/actors/{actor_id}/scrape": {
      "post": {
        "tags": [
          "actors"
        ],
        "summary": "Scrape Actor",
        "description": "提交演员人物元数据刮削任务.",
        "operationId": "scrape_actor",
        "parameters": [
          {
            "name": "actor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Actor Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ActorScrapeRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/facets/{kind}/{facet_id}": {
      "get": {
        "tags": [
          "facets"
        ],
        "summary": "Get Facet",
        "operationId": "get_facet",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FacetKind"
            }
          },
          {
            "name": "facet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Facet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "facets"
        ],
        "summary": "Rename Facet",
        "description": "重命名分类实体. 新名称与另一实体冲突时返回 409 (建议改用合并).",
        "operationId": "rename_facet",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FacetKind"
            }
          },
          {
            "name": "facet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Facet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacetRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "facets"
        ],
        "summary": "Delete Facet",
        "description": "删除分类. 爬取侧写入黑名单并从 Metadata 真值剔除; user_tag 硬删.",
        "operationId": "delete_facet",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FacetKind"
            }
          },
          {
            "name": "facet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Facet Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorGender": {
        "type": "string",
        "enum": [
          "female",
          "male",
          "unknown"
        ],
        "title": "ActorGender",
        "description": "演员性别 - 用于展示与按站裁剪刮削源."
      },
      "ActorListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActorResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ActorListResponse"
      },
      "ActorResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases",
            "description": "别名行 (保序; 不含展示名)"
          },
          "gender": {
            "$ref": "#/components/schemas/ActorGender",
            "default": "unknown"
          },
          "birthday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday"
          },
          "birthplace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthplace"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "bust": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bust"
          },
          "waist": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waist"
          },
          "hip": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hip"
          },
          "cup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cup"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "image_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Image Urls"
          },
          "provider_ids": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Provider Ids"
          },
          "source_urls": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Source Urls"
          },
          "field_sources": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Field Sources"
          },
          "raw": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Raw"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ActorResponse",
        "description": "详情填全量; 列表 (`GET /actors`) 只填卡片/表格字段, 简介/别名/源字典/raw 为空."
      },
      "ActorScrapeRequest": {
        "properties": {
          "use_cache": {
            "items": {
              "$ref": "#/components/schemas/CacheKind"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Use Cache",
            "description": "启用的缓存种类 (metadata: 复用 Actor.raw; trans: 预留译文). 空集 = 全部强制刷新"
          }
        },
        "type": "object",
        "title": "ActorScrapeRequest",
        "description": "POST /actors/{id}/scrape 请求体 - 与影片刮削同型 use_cache."
      },
...
      "ActorSortField": {
        "type": "string",
        "enum": [
          "name",
          "count",
          "updated_at",
          "has_image",
          "birthday",
          "height",
          "bust",
          "waist",
          "hip",
          "cup"
        ],
        "title": "ActorSortField",
        "description": "演员浏览列表排序字段."
      },
      "ActorUpdateRequest": {
        "properties": {
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActorGender"
              },
              {
                "type": "null"
              }
            ]
          },
          "birthday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday"
          },
          "birthplace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthplace"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "bust": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bust"
          },
          "waist": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waist"
          },
          "hip": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hip"
          },
          "cup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cup"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "image_urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Urls"
          },
          "provider_ids": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Ids"
          },
          "source_urls": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Urls"
          },
          "aliases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aliases",
            "description": "别名行 (保序), 整表替换"
          }
        },
        "type": "object",
        "title": "ActorUpdateRequest"
      },
...
      "CacheKind": {
        "type": "string",
        "enum": [
          "metadata",
          "trans"
        ],
        "title": "CacheKind",
        "description": "刮削可复用的缓存种类. use_cache 为其集合: 含某项 = 该缓存生效, 不含 = 强制刷新该项."
      },
...
      "FacetKind": {
        "type": "string",
        "enum": [
          "actor",
          "director",
          "tag",
          "studio",
          "publisher",
          "series",
          "user_tag"
        ],
        "title": "FacetKind",
        "description": "分类索引种类 - 爬取侧目录 + 用户 tag."
      },
...
      "FacetRenameRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "新名称"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FacetRenameRequest"
      },
      "FacetResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "count"
        ],
        "title": "FacetResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
...
    }
  }
}